# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Mist Automation & Backup API
  description: |
    Comprehensive web application for automating Juniper Mist operations and managing configuration backups.
    
    ## Authentication
    This API uses JWT Bearer token authentication. To authenticate:
    1. Call `POST /api/v1/auth/login` with your credentials
    2. Use the returned `access_token` in the `Authorization` header as `Bearer <token>`
    
    ## First-Time Setup (Onboarding)
    When the system has no users, use `POST /api/v1/auth/onboard` to create the first admin user.
    
    ## Rate Limiting
    API requests are rate-limited per user. Please handle 429 responses appropriately.
    Login attempts are rate-limited to 5 attempts per 5 minutes per IP/email combination.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: API v1
...
security:
  - BearerAuth: []
paths:
...
  /admin/settings:
    get:
      tags:
        - Admin
      summary: Get system settings
      description: Get system configuration settings (admin only)
      operationId: getSystemSettings
      responses:
        '200':
          description: System settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettingsResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Admin
      summary: Update system settings
      description: Update system configuration settings (admin only)
      operationId: updateSystemSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettingsUpdate'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessageResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/logs:
    get:
      tags:
        - Admin
      summary: Get audit logs
      description: Get system audit logs (admin only)
      operationId: getAuditLogs
      parameters:
        - name: skip
          in: query
          description: Number of logs to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of logs to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: event_type
          in: query
          description: Filter by event type
          schema:
            type: string
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: string
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/stats:
    get:
      tags:
        - Admin
      summary: Get system stats
      description: Get system statistics and metrics (admin only)
      operationId: getSystemStats
      responses:
        '200':
          description: System statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatsResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/mist/test-connection:
    post:
      tags:
        - Admin
      summary: Test Mist connection
      description: |
        Test connection to Mist API (admin only).
        Accepts optional mist_api_token, mist_org_id, and mist_cloud_region in the request body.
        When provided, the connection is tested with those values directly (no save required).
        Missing fields fall back to saved system configuration.
      operationId: testMistConnection
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MistConnectionTestRequest'
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionTestResponse'
        '400':
          description: Mist API token not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/mist/sites:
    get:
      tags:
        - Admin
      summary: List Mist sites
      description: List sites from the configured Mist organization
      operationId: listMistSites
      responses:
        '200':
          description: List of Mist sites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MistSitesResponse'
        '400':
          description: Mist API not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Failed to fetch from Mist API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/mist/object-types:
    get:
      tags:
        - Admin
      summary: List Mist object types
      description: Return all supported Mist object types for frontend dropdowns
      operationId: listMistObjectTypes
      responses:
        '200':
          description: List of Mist object types
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MistObjectTypesResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/mist/objects:
    get:
      tags:
        - Admin
      summary: List Mist objects
      description: List objects of a given type from Mist organization. Uses the object registry for consistent API calls.
      operationId: listMistObjects
      parameters:
        - name: object_type
          in: query
          required: true
          description: Object type in 'org:key' or 'site:key' format
          schema:
            type: string
        - name: site_id
          in: query
          description: Site ID (required for site-level objects)
          schema:
            type: string
      responses:
        '200':
          description: List of Mist objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MistObjectsResponse'
        '400':
          description: Invalid request or Mist API not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Failed to fetch from Mist API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/workers/status:
    get:
      tags:
        - Admin
      summary: Get worker status
      description: Get status of background workers (admin only)
      operationId: getWorkerStatus
      responses:
        '200':
          description: Worker status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerStatusResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════════════
  # Dashboard Endpoints
  # ═══════════════════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
    StatusMessageResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
        message:
          type: string

    # ─────────────────────────────────────────────────────────────────────────
    # Health
    # ─────────────────────────────────────────────────────────────────────────
...
    SystemSettingsResponse:
      type: object
      required:
        - mist_api_token_set
        - max_concurrent_workflows
        - workflow_default_timeout
      properties:
        mist_api_token_set:
          type: boolean
          description: Whether Mist API token is configured
        mist_org_id:
          type:
            - string
            - 'null'
          description: Mist organization ID
        mist_cloud_region:
          type:
            - string
            - 'null'
          description: Mist cloud region
        webhook_secret:
          type:
            - string
            - 'null'
          description: Webhook verification secret (decrypted)
        smee_enabled:
          type: boolean
          description: Whether Smee.io is enabled
        smee_channel_url:
          type:
            - string
            - 'null'
          description: Smee.io channel URL
        max_concurrent_workflows:
          type: integer
          description: Maximum concurrent workflows
        workflow_default_timeout:
          type: integer
          description: Default workflow timeout in seconds
        min_password_length:
          type: integer
          description: Minimum password length
        require_uppercase:
          type: boolean
        require_lowercase:
          type: boolean
        require_digits:
          type: boolean
        require_special_chars:
          type: boolean
        session_timeout_hours:
          type: integer
          description: Session timeout in hours
        max_concurrent_sessions:
          type: integer
          description: Maximum concurrent sessions per user
        backup_enabled:
          type: boolean
          description: Whether scheduled backups are enabled
        backup_full_schedule_cron:
          type:
            - string
            - 'null'
          description: Cron expression for full backups
        backup_retention_days:
          type: integer
          description: Backup retention period in days
        backup_git_enabled:
          type: boolean
          description: Whether Git backup is enabled
        backup_git_repo_url:
          type:
            - string
            - 'null'
          description: Git repository URL
        backup_git_branch:
          type:
            - string
            - 'null'
          description: Git branch name
        backup_git_author_name:
          type:
            - string
            - 'null'
          description: Git commit author name
        backup_git_author_email:
          type:
            - string
            - 'null'
          description: Git commit author email
        slack_webhook_url:
          type:
            - string
            - 'null'
          description: Slack webhook URL for notifications
        servicenow_instance_url:
          type:
            - string
            - 'null'
          description: ServiceNow instance URL
        servicenow_username:
          type:
            - string
            - 'null'
          description: ServiceNow username
        servicenow_password_set:
          type: boolean
          description: Whether ServiceNow password is configured
        pagerduty_api_key_set:
          type: boolean
          description: Whether PagerDuty API key is configured
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
          description: Last update timestamp
    SystemSettingsUpdate:
      type: object
      properties:
        mist_api_token:
          type: string
          description: Mist API token
        mist_org_id:
          type: string
          description: Mist organization ID
        mist_cloud_region:
          type: string
          description: Mist cloud region
        webhook_secret:
          type: string
          description: Webhook verification secret
        smee_enabled:
          type: boolean
          description: Enable Smee.io
        smee_channel_url:
          type: string
          description: Smee.io channel URL
        max_concurrent_workflows:
          type: integer
          description: Maximum concurrent workflows
        workflow_default_timeout:
          type: integer
          description: Default workflow timeout
        min_password_length:
          type: integer
        require_uppercase:
          type: boolean
        require_lowercase:
          type: boolean
        require_digits:
          type: boolean
        require_special_chars:
          type: boolean
        session_timeout_hours:
          type: integer
        max_concurrent_sessions:
          type: integer
        backup_enabled:
          type: boolean
        backup_full_schedule_cron:
          type: string
        backup_retention_days:
          type: integer
        backup_git_enabled:
          type: boolean
        backup_git_repo_url:
          type: string
        backup_git_branch:
          type: string
        backup_git_author_name:
          type: string
        backup_git_author_email:
          type: string
        slack_webhook_url:
          type: string
        servicenow_instance_url:
          type: string
        servicenow_username:
          type: string
        servicenow_password:
          type: string
        pagerduty_api_key:
          type: string
    MistConnectionTestRequest:
      type: object
      properties:
        mist_api_token:
          type: string
          description: Mist API token to test
        mist_org_id:
          type: string
          description: Mist organization ID
        mist_cloud_region:
          type: string
          description: Mist cloud region
...
    AuditLogListResponse:
      type: object
      required:
        - logs
        - total
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogResponse'
          description: List of audit logs
        total:
          type: integer
          description: Total number of logs
    SystemStatsResponse:
      type: object
      required:
        - workflows
        - executions
        - backups
        - webhooks
        - users
      properties:
        workflows:
          type: object
          properties:
            total:
              type: integer
            enabled:
              type: integer
            draft:
              type: integer
        executions:
          type: object
          properties:
            total:
              type: integer
            pending:
              type: integer
            running:
              type: integer
            succeeded:
              type: integer
            failed:
              type: integer
        backups:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
            pending:
              type: integer
            failed:
              type: integer
        webhooks:
          type: object
          properties:
            total:
              type: integer
            processed:
              type: integer
            pending:
              type: integer
        users:
          type: object
          properties:
            total:
              type: integer
            active:
              type: integer
            admins:
              type: integer
    ConnectionTestResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [connected, failed]
          description: Connection status
        error:
          type:
            - string
            - 'null'
          description: Error message if failed
...
    MistSitesResponse:
      type: object
      required:
        - sites
      properties:
        sites:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Site ID
              name:
                type: string
                description: Site name
    MistObjectTypesResponse:
      type: object
      required:
        - object_types
      properties:
        object_types:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
                description: Object type value (e.g., 'org:wlans')
              label:
                type: string
                description: Human-readable label
              scope:
                type: string
                enum: [org, site]
                description: Object scope
    MistObjectsResponse:
      type: object
      required:
        - objects
      properties:
        objects:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Object ID
              name:
                type: string
                description: Object name
              type:
                type: string
                description: Object type

    # ─────────────────────────────────────────────────────────────────────────
    # Device Utils Schemas
    # ─────────────────────────────────────────────────────────────────────────
...